+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fixed
   - Fix type parameter generation
   - Fix handling of `class Foo::Bar`
+- Misc
+  - Change order of building `RbsMacros::Environment`
 
 ## [0.1.0] - 2024-02-29